anomaly detection service

+++ api/src/main/java/com/devtools/task_time_tracker/controller/TaskController.java
@@ -22,7 +22,7 @@ public class TaskController {
     public ResponseEntity<TaskModel> createTask(
             @RequestParam String description,
             @RequestParam String name,
-            @RequestParam Integer storyPoints,
+            @RequestParam Long storyPoints,
             @RequestParam UUID projectId
     ) {
         TaskModel task = taskService.createTask(description, name, storyPoints, projectId);
